Need to be able to set the IIS application pool and website defaults settings.
The objects for an application pool and the application pool defaults are the same. So are the objects for a website
and the website defaults. If we make code that gets application pools and websites also support getting defaults, any
set function we create now and in the future, don't really need to be aware if they are modifying defaults or an actual
applcation pool or website.

* Updated Get-CIisWebsite to return the site defaults settings, using the
`Microsoft.Web.Administration.ServerManager.SiteDefaults` property.
* Updated Get-CIisAppPool to return the ApplicationPoolDefaults settings, using the
`Microsoft.Web.Administration.ServerManager.ApplicationPoolDefaults` property.
* Noticed that using [switch] for boolean settings means they're always set to false. Use [bool] instead, that way
their value only changes if someone explicitly passes a value.
* Updated Set-CIisWebsiteLogFile to configure website defaults settings.
* Updated Set-CIisAppPoolCpu and Set-CIisAppPoolRecyclingPeriodicRestart to configure application pool defaults
settings.