Creating function Set-CIisWebsite for configuring website settings.
These are the settings on the <site> element in the applicationHost.config file, which are currently only
`serverStartMode` and `Id`. The `Id` parameter can only be set and changed. It can never be unset/removed.

Updated `Install-CIisWebsite` to match the parameters of `Set-CIisWebsite` so users don't need to call two functions to
install and configure a new website, and to make it easier to call `Set-CIisWebsite` from `Install-CIisWebsite`.